backend/requirements/base.txt
Django==5.2.12
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-decouple==3.8
whitenoise==6.12.0
